requirements.txt
mmh3==3.0.0
requests==2.28.1
netlas==0.4.0
PyYAML==6.0
shodan==1.28.0